+++ packages/grafana-data/src/dataframe/index.ts
@@ -6,3 +6,4 @@ export * from './processDataFrame';
 export * from './dimensions';
 export * from './ArrowDataFrame';
 export * from './ArrayDataFrame';
+export * from './toDataFrame';
