+++ packages/grafana-data/src/types/data.ts
@@ -2,8 +2,6 @@ import { FieldConfig } from './dataFrame';
 import { DataTransformerConfig } from './transformations';
 import { ApplyFieldOverrideOptions } from './fieldOverrides';
 
-export type KeyValue<T = any> = { [s: string]: T };
-
 /**
  * Represent panel data loading state.
  */
